sw_backend_server/python-server/requirements.txt
anyio==3.5.0
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
deprecation==2.1.0
gotrue==0.5.0
h11==0.12.0
httpcore==0.14.7
httpx==0.21.3
idna==3.3
Localization==0.1.7
numpy==1.22.3
packaging==21.3
postgrest-py==0.9.2
pydantic==1.9.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
realtime==0.0.4
redis==4.2.2
rfc3986==1.5.0
scipy==1.8.0
Shapely==1.8.1.post1
six==1.16.0
sniffio==1.2.0
supabase==0.5.3
typing_extensions==4.2.0
websockets==9.1
wrapt==1.14.0
